--- app/src/main/java/com/psi/thrifter/data_trending.kt
@@ -1,7 +1,0 @@
-package com.example.ujicoba
-
-data class data_trending(
-    var image: Int,
-    var nama: String,
-    var harga: String
-)
