+++ app/src/main/java/com/psi/thrifter/kategori/MainActivity_kategori.kt
@@ -1,43 +1,45 @@
-package com.example.ujicoba32
+package com.psi.thrifter.kategori
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import androidx.appcompat.app.ActionBar
-import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.ujicoba.data_trending
-import com.example.ujicoba.adaptertrendingpage
 import com.psi.thrifter.R
 
-class MainActivity : AppCompatActivity() {
-    lateinit var main: RecyclerView
+class MainActivity_kategori : AppCompatActivity() {
+    lateinit var main : RecyclerView
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_trending)
-        var actionBar = supportActionBar
+        setContentView(R.layout.activity_kategori)
+        var actionBar = getSupportActionBar()
 
         // showing the back button in action bar
         if (actionBar != null) {
-            actionBar.displayOptions = ActionBar.DISPLAY_SHOW_CUSTOM
-            actionBar.setCustomView(R.layout.actionbar_trending)
-            actionBar.setHomeAsUpIndicator(R.drawable.back_button)
+            actionBar.setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM)
+            actionBar.setCustomView(R.layout.actionbar_kategori)
+            actionBar.setHomeAsUpIndicator(R.drawable.back_button);
             actionBar.setDisplayHomeAsUpEnabled(true)
             actionBar.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
         }
 
 
         val todoVM = listOf(
-            //database
+            data_kategori(R.drawable.kategori_baju,"Baju"),
+            data_kategori(R.drawable.kategori_celana,"Celana"),
+            data_kategori(R.drawable.kategori_sepatu,"topi"),
+            data_kategori(R.drawable.kategori_topi,"Topi"),
+
         )
 
         main = findViewById(R.id.recyclerSemuaKategoriPage)
         main.apply {
-            layoutManager = GridLayoutManager(this@MainActivity, 2)
-            adapter = adaptertrendingpage(todoVM)
+            layoutManager = GridLayoutManager(this@MainActivity_kategori, 2)
+            adapter = adapterKategori(todoVM)
         }
 
     }
