+++ app/src/main/java/com/psi/thrifter/kategori/SearchFragment.kt
@@ -1,32 +1,22 @@
-package com.psi.thrifter
+package com.psi.thrifter.kategori
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.psi.thrifter.R
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [SearchFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class SearchFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
+    lateinit var main: RecyclerView
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+
         }
+
     }
 
     override fun onCreateView(
@@ -34,8 +24,27 @@ class SearchFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_search, container, false)
+        val root = inflater.inflate(R.layout.fragment_search,container,false)
+
+
+        val todoVM = listOf(
+            data_kategori(R.drawable.kategori_baju,"Baju"),
+            data_kategori(R.drawable.kategori_celana,"Celana"),
+            data_kategori(R.drawable.kategori_sepatu,"topi"),
+            data_kategori(R.drawable.kategori_topi,"Topi")
+
+            )
+
+        main = root.findViewById(R.id.recyclerSemuaKategoriPage)
+        main.apply {
+            layoutManager = GridLayoutManager(this.context, 2)
+            adapter = adapterKategori(todoVM)
+        }
+
+        return root
     }
+
+
 
     companion object {
         /**
@@ -51,8 +60,7 @@ class SearchFragment : Fragment() {
         fun newInstance(param1: String, param2: String) =
             SearchFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+
                 }
             }
     }
