+++ app/src/main/java/com/psi/thrifter/kategori/data_kategori.kt
@@ -1,4 +1,4 @@
-package com.example.ujicoba
+package com.psi.thrifter.kategori
 
 data class data_kategori(
     val image: Int,
